Renamed from artifacts to posts removed gutter as padding handled in blocks component
